Create CS315Lab_5.cpp
This program first takes in a string and makes it a list of tokens. Then, the program converts this list into a nested list structure of integers. The program can then print the contents of the array along with the depth of each element. The program can also flatten the structure so that each element is of the same depth or change one of  elements to a different value.